fix: more tests
