+++ package.json
@@ -18,16 +18,16 @@
 		"@sentry/sveltekit": "^8.55.0",
 		"@sveltejs/adapter-static": "^3.0.8",
 		"@sveltejs/kit": "^2.16.0",
-		"@testing-library/svelte": "^5.2.4",
-		"@thisux/sveltednd": "^0.0.16",
+		"@testing-library/svelte": "^5.2.8",
+		"@thisux/sveltednd": "0.0.20",
 		"@total-typescript/ts-reset": "^0.5.1",
 		"@types/node": "^18.19.59",
 		"@types/pako": "^2.0.3",
 		"@types/wicg-file-system-access": "^2020.9.6",
-		"@vitest/web-worker": "^2.1.8",
+		"@vitest/web-worker": "^3.2.4",
 		"comlink": "^4.4.2",
 		"fast-xml-parser": "^4.5.0",
-		"happy-dom": "^15.11.6",
+		"happy-dom": "^18.0.1",
 		"idb": "^7.1.1",
 		"image-dimensions": "^2.3.0",
 		"pako": "^2.1.0",
@@ -40,7 +40,7 @@
 		"tsx": "^4.19.1",
 		"typescript": "^5.0.0",
 		"vite": "^6.0.7",
-		"vitest": "^2.1.4",
+		"vitest": "^3.2.4",
 		"vitest-indexeddb": "^0.0.0"
 	},
 	"optionalDependencies": {
