+++ tests/e2e.svelte.test.ts
@@ -1,0 +1,366 @@
+import { readFileSync } from "node:fs";
+import { readFile } from "node:fs/promises";
+import { fireEvent, render, within } from "@testing-library/svelte";
+import { flushSync, tick } from "svelte";
+import { beforeAll, beforeEach, describe, expect, it, vi } from "vitest";
+import { setup as mockIDB } from "vitest-indexeddb";
+import {
+	SaveManager,
+	setSaveManagerContext,
+} from "../src/lib/SaveManager.svelte";
+import craftingPage from "../src/routes/(edit)/(list)/crafting/+page.svelte";
+import appearancePage from "../src/routes/(edit)/appearance/+page.svelte";
+import characterPage from "../src/routes/(edit)/character/+page.svelte";
+import editorPage from "../src/routes/(edit)/inventory/+page.svelte";
+import relationshipsPage from "../src/routes/(edit)/relationships/+page.svelte";
+import { Gender } from "../codegen/save";
+import bundlesPage from "../src/routes/(edit)/bundles/+page.svelte";
+
+describe("Save Manager Integration Tests", () => {
+	// Wrap in a root effect so we can initialize the SaveManager
+	// $effect.root(() => {
+	mockIDB(); // Mock IndexedDB for testing
+	let saveManager: SaveManager;
+
+	const renderWithContext = ((component, options) =>
+		// @ts-ignore
+		render(component, {
+			context: setSaveManagerContext(saveManager),
+			...options,
+		})) as typeof render;
+
+	// Mock the alert function
+	beforeAll(async () => {
+		vi.spyOn(window, "alert").mockImplementation((m) =>
+			console.log(`[alert]: ${m}`),
+		);
+	});
+
+	// Reset the save manager before each test
+	beforeEach(async () => {
+		let promise: Promise<void> = Promise.resolve();
+		const file = await readFile("tests/TestSave");
+
+		$effect.root(() => {
+			saveManager = new SaveManager();
+			promise = saveManager.import(new File([file], "TestSave"));
+		});
+
+		// I don't even know what's going on at this point
+		await promise;
+
+		flushSync();
+	});
+
+	describe("Community Bundles", () => {
+		it("should list bundles and toggle a required item", async () => {
+			const page = renderWithContext(bundlesPage);
+			// Ensure we are in community bundles mode (not Joja)
+			const membership = page.getByTestId(
+				"joja-membership-checkbox",
+			) as HTMLInputElement;
+			if (membership.checked) {
+				await fireEvent.click(membership); // disable
+			}
+
+			const bundleCards = page.getAllByTestId("bundle-card");
+			expect(bundleCards.length).toBeGreaterThan(0);
+			// Expand first bundle
+			const firstCard = bundleCards[0] as HTMLElement;
+			const firstDetails = firstCard.querySelector("details");
+			if (firstDetails) firstDetails.open = true;
+			await tick();
+
+			const requiredContainer = firstCard.querySelector(
+				'[data-testid="bundle-required-items"]',
+			);
+			expect(requiredContainer).not.toBeNull();
+			if (!requiredContainer) return;
+			const firstItem = requiredContainer.querySelector(
+				'[data-testid="bundle-required-item"] input[type="checkbox"]',
+			) as HTMLInputElement | null;
+			if (firstItem) {
+				const prev = firstItem.checked;
+				await fireEvent.click(firstItem);
+				expect(firstItem.checked).not.toBe(prev);
+			}
+		});
+
+		it("should toggle Joja membership and set a project", async () => {
+			const page = renderWithContext(bundlesPage);
+			const membership = page.getByTestId(
+				"joja-membership-checkbox",
+			) as HTMLInputElement;
+			const initial = membership.checked;
+			await fireEvent.click(membership); // toggle
+			expect(membership.checked).not.toBe(initial);
+
+			if (membership.checked) {
+				// now in Joja mode
+				const projectRows = page.getAllByTestId("joja-project-row");
+				expect(projectRows.length).toBeGreaterThan(0);
+				const firstRow = projectRows[0] as HTMLElement;
+				const firstCheckbox = firstRow.querySelector(
+					'[data-testid="joja-project-checkbox"]',
+				) as HTMLInputElement | null;
+				if (firstCheckbox) {
+					const prev = firstCheckbox.checked;
+					await fireEvent.click(firstCheckbox);
+					expect(firstCheckbox.checked).not.toBe(prev);
+				}
+			}
+		});
+	});
+
+	describe("Inventory Management", () => {
+		it("should create and edit items", async () => {
+			const page = renderWithContext(editorPage);
+			const slot = page.getByTestId("item-9");
+			await fireEvent.click(slot);
+			await tick();
+
+			const input = page.getByTestId("item-name") as HTMLInputElement;
+			// New UI: type to filter then select from dropdown suggestions
+			await fireEvent.input(input, { target: { value: "Leek" } });
+			await tick(); // wait for derived values to update
+			// Click the suggestion button containing the item name
+			await fireEvent.click(page.getByText("Leek"));
+
+			expect(saveManager.save?.player.inventory.items.get(9)?.name).toBe(
+				"Leek",
+			);
+		});
+
+		it("should modify item quality", async () => {
+			const page = renderWithContext(editorPage);
+			await fireEvent.click(page.getByTestId("item-4"));
+			await tick();
+
+			for (const quality of [0, 1, 2, 4]) {
+				await fireEvent.click(page.getByTestId(`quality-${quality}`));
+				expect(saveManager.save?.player.inventory.items.get(4)?.quality).toBe(
+					quality,
+				);
+			}
+		});
+
+		// it("should support drag and drop operations", async () => {
+		// 	const page = renderWithContext(editorPage);
+		// 	const draggable = page.getByTestId("draggable-4");
+		// 	const destination = page.getByTestId("slot-6");
+
+		// 	await fireEvent.dragStart(draggable);
+		// 	await fireEvent.dragEnter(destination);
+		// 	await fireEvent.dragOver(destination);
+		// 	await fireEvent.drop(destination);
+		// 	await fireEvent.dragEnd(draggable);
+
+		// 	expect(saveManager.save.player.inventory.getItem(6).name).toBe(
+		// 		"Parsnip",
+		// 	);
+		// 	expect(saveManager.save.player.inventory.getItem(4)).toBeUndefined();
+		// });
+	});
+
+	describe("Character Customization", () => {
+		it("should modify character appearance", () => {
+			$effect.root(() => {
+				const page = renderWithContext(appearancePage);
+				const changes = [
+					["name", "Test2"],
+					["farmName", "Test2"],
+					["favoriteThing", "Integration Tests"],
+					["skin", 1],
+					["hairstyle", 68],
+					["accessory", 26],
+				] as const;
+
+				for (const [name, value] of changes) {
+					const input = page.getByTestId(
+						`appearance-${name}`,
+					) as HTMLInputElement;
+					fireEvent.input(input, { target: { value } });
+					expect(saveManager.save?.player[name]).toBe(value);
+				}
+			});
+		});
+
+		it("should update amount and delete an item", async () => {
+			const page = renderWithContext(editorPage);
+			// Select slot 9 (empty or existing) then create an item first
+			await fireEvent.click(page.getByTestId("item-9"));
+			await tick();
+			const input = page.getByTestId("item-name") as HTMLInputElement;
+			await fireEvent.input(input, { target: { value: "Leek" } });
+			await tick();
+			await fireEvent.click(page.getByText("Leek"));
+			await tick();
+
+			// Change the amount property if present
+			const amountInput = page.queryByTestId(
+				"property-amount",
+			) as HTMLInputElement | null;
+			if (amountInput) {
+				await fireEvent.input(amountInput, { target: { value: "5" } });
+				expect(saveManager.save?.player.inventory.items.get(9)?.amount).toBe(5);
+			}
+
+			// Delete button (trash emoji)
+			const deleteBtn = page.getByRole("button", { name: "🗑️" });
+			await fireEvent.click(deleteBtn);
+			await tick();
+			expect(saveManager.save?.player.inventory.items.get(9)).toBeUndefined();
+		});
+
+		it("should change gender and hair color", async () => {
+			const page = renderWithContext(appearancePage);
+			// Toggle gender (click opposite of current)
+			const currentGender = saveManager.save?.player.gender;
+			const target = currentGender === Gender.Male ? "female" : "male";
+			const radio = page.getByDisplayValue(target) as HTMLInputElement;
+			await fireEvent.click(radio);
+			if (saveManager.save) {
+				expect(saveManager.save.player.gender).not.toBe(currentGender);
+			}
+
+			// Change hair color
+			const hairColor = page.getByTestId(
+				"appearance-hairColor",
+			) as HTMLInputElement;
+			await fireEvent.change(hairColor, { target: { value: "#123456" } });
+			if (saveManager.save) {
+				expect(saveManager.save.player.hairColor.toHex().toLowerCase()).toBe(
+					"#123456",
+				);
+			}
+		});
+	});
+	describe("Skills and Crafting", () => {
+		it("should modify experience values", async () => {
+			const page = renderWithContext(characterPage);
+			const skills = [
+				["farming", 0],
+				["mining", 3],
+				["combat", 4],
+				["foraging", 2],
+				["fishing", 1],
+			] as const;
+
+			for (const [name, value] of skills) {
+				const input = page.getByTestId(`skills-${name}`) as HTMLInputElement;
+				await fireEvent.input(input, { target: { value } });
+				expect(saveManager.save?.player.skills.raw[value]).toBe(value);
+			}
+		});
+
+		it("should modify crafting recipes", async () => {
+			const page = renderWithContext(craftingPage);
+			expect(saveManager.save?.player.craftingRecipes.get("Keg")).toBeNull();
+			await fireEvent.click(page.getByTestId("recipe-keg"));
+			expect(
+				saveManager.save?.player.craftingRecipes.get("Keg"),
+			).not.toBeNull();
+		});
+
+		it("should toggle recipe (unlock & lock)", async () => {
+			const page = renderWithContext(craftingPage);
+			const recipe = page.getByTestId("recipe-keg");
+			// Unlock
+			await fireEvent.click(recipe);
+			expect(
+				saveManager.save?.player.craftingRecipes.get("Keg"),
+			).not.toBeNull();
+			// Lock again
+			await fireEvent.click(recipe);
+			expect(saveManager.save?.player.craftingRecipes.get("Keg")).toBeNull();
+		});
+
+		it("should modify character stats", async () => {
+			const page = renderWithContext(characterPage);
+			if (!saveManager.save) return;
+			const save = saveManager.save;
+			const setNumber = async (
+				labelRegex: RegExp,
+				value: number,
+				setter: (v: number) => void,
+				getter: () => number,
+			) => {
+				const input = page.getByLabelText(labelRegex) as HTMLInputElement;
+				await fireEvent.input(input, { target: { value: String(value) } });
+				setter(value); // ensure reactivity sync in case
+				expect(getter()).toBe(value);
+			};
+
+			await setNumber(
+				/Health/,
+				123,
+				(v) => {
+					save.player.maxHealth = v;
+				},
+				() => save.player.maxHealth,
+			);
+			await setNumber(
+				/Stamina/,
+				234,
+				(v) => {
+					save.player.maxStamina = v;
+				},
+				() => save.player.maxStamina,
+			);
+			await setNumber(
+				/Golden Walnuts/,
+				12,
+				(v) => {
+					save.goldenWalnuts = v;
+				},
+				() => save.goldenWalnuts ?? 0,
+			);
+			await setNumber(
+				/Hay/,
+				50,
+				(v) => {
+					if (save.farm) save.farm.piecesOfHay = v;
+				},
+				() => save.farm?.piecesOfHay ?? 0,
+			);
+			await setNumber(
+				/Deepest Mine/,
+				42,
+				(v) => {
+					save.deepestMineLevel = v;
+				},
+				() => save.deepestMineLevel,
+			);
+		});
+	});
+
+	describe("Relationships", () => {
+		it("should update friendship hearts", async () => {
+			const page = renderWithContext(relationshipsPage);
+			if (!saveManager.save) return;
+			const rows = page.getAllByTestId("friendship-row");
+			expect(rows.length).toBeGreaterThan(0);
+			const firstRow = rows[0] as HTMLElement;
+			const nameEl = within(firstRow).getByTestId("friendship-name");
+			const name = nameEl.textContent?.trim();
+			expect(name).toBeTruthy();
+			if (!name) return;
+			const friendship = saveManager.save.player.friendships.get(name);
+			expect(friendship).toBeTruthy();
+			if (!friendship) return;
+			const buttons = within(firstRow).getAllByTestId("friendship-heart");
+			expect(buttons.length).toBeGreaterThan(0);
+			const initialHearts = friendship.hearts;
+			let targetHearts = 1;
+			if (initialHearts === 1) targetHearts = 2;
+			else if (initialHearts > 2) targetHearts = 1;
+			const targetIndex = targetHearts - 1;
+			if (buttons[targetIndex]) {
+				await fireEvent.click(buttons[targetIndex]);
+				expect(friendship.hearts).toBe(targetHearts);
+				expect(friendship.points).toBe(friendship.hearts * 250);
+			}
+		});
+	});
+	// });
+});
