+++ tsconfig.json
@@ -9,7 +9,10 @@
 		"skipLibCheck": true,
 		"sourceMap": true,
 		"strict": true,
-		"typeRoots": ["./node_modules/@types", "./node_modules/@total-typescript"],
+		"typeRoots": [
+			"./node_modules/@types",
+			"./node_modules/@total-typescript"
+		],
 		"noUncheckedIndexedAccess": true
 	},
 	// Path aliases are handled by https://kit.svelte.dev/docs/configuration#alias
@@ -24,6 +27,6 @@
 		"src/service-worker/**/*.ts",
 		"src/service-worker.d.ts",
 		"src/service-worker/**/*.d.ts",
-		"codegen/dump.tss"
+		"codegen/dump.ts"
 	]
 }
